# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TorqueLane API",
        "version": "0.1.0",
        "description": "TorqueLane API. Errors always use the envelope { \"error\": { \"code\", \"message\", \"details\"? } }; clients branch on `code`. Records outside the tenant scope of the caller answer 404 not_found, exactly like records that do not exist."
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Relative to the environment host, e.g. https://api-staging.<domain>"
        }
    ],
    "paths": {
...
        "/bays/{bay}": {
            "get": {
                "operationId": "bays.show",
                "summary": "Show a bay",
                "tags": [
                    "Bay"
                ],
                "parameters": [
                    {
                        "name": "bay",
                        "in": "path",
                        "required": true,
                        "description": "The bay ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BayResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BayResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            },
            "put": {
                "operationId": "bays.update",
                "summary": "Update a bay",
                "tags": [
                    "Bay"
                ],
                "parameters": [
                    {
                        "name": "bay",
                        "in": "path",
                        "required": true,
                        "description": "The bay ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveBayRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`BayResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BayResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error.validation"
                    }
                }
            },
            "delete": {
                "operationId": "bays.destroy",
                "description": "A technician's home bay is a 409.",
                "summary": "Delete a bay",
                "tags": [
                    "Bay"
                ],
                "parameters": [
                    {
                        "name": "bay",
                        "in": "path",
                        "required": true,
                        "description": "The bay ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
        "/branches/{branch}": {
            "get": {
                "operationId": "branches.show",
                "summary": "Show a branch",
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "description": "The branch ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BranchResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BranchResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            },
            "put": {
                "operationId": "branches.update",
                "description": "Needs `settings:manage` and access to the branch.",
                "summary": "Update a branch",
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "description": "The branch ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveBranchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`BranchResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BranchResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error.validation"
                    }
                }
            },
            "delete": {
                "operationId": "branches.destroy",
                "description": "Needs `organization:manage`. A branch with bays, technicians, pinned\nstaff or document series is a 409: mark it inactive instead.",
                "summary": "Delete an empty branch",
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "description": "The branch ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
        "/customer-accounts/{customerAccount}/contacts/{contact}": {
            "get": {
                "operationId": "customer-accounts.contacts.show",
                "summary": "Show a contact",
                "tags": [
                    "Contact"
                ],
                "parameters": [
                    {
                        "name": "customerAccount",
                        "in": "path",
                        "required": true,
                        "description": "The customer account ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "description": "The contact ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            },
            "put": {
                "operationId": "customer-accounts.contacts.update",
                "summary": "Update a contact",
                "tags": [
                    "Contact"
                ],
                "parameters": [
                    {
                        "name": "customerAccount",
                        "in": "path",
                        "required": true,
                        "description": "The customer account ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "description": "The contact ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ContactResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ContactResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error.validation"
                    }
                }
            },
            "delete": {
                "operationId": "customer-accounts.contacts.destroy",
                "description": "A contact with consent decisions on record cannot be deleted (409).",
                "summary": "Delete a contact",
                "tags": [
                    "Contact"
                ],
                "parameters": [
                    {
                        "name": "customerAccount",
                        "in": "path",
                        "required": true,
                        "description": "The customer account ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "description": "The contact ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
        "/documents/{document}": {
            "get": {
                "operationId": "documents.show",
                "summary": "Show a document",
                "tags": [
                    "Document"
                ],
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "required": true,
                        "description": "The document ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`DocumentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DocumentResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            },
            "delete": {
                "operationId": "documents.destroy",
                "description": "`document:delete`. The file is removed after the row's deletion commits.",
                "summary": "Delete a document",
                "tags": [
                    "Document"
                ],
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "required": true,
                        "description": "The document ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
        "/invitations/{invitation}": {
            "delete": {
                "operationId": "invitations.destroy",
                "summary": "Revoke an invitation",
                "tags": [
                    "Invitation"
                ],
                "parameters": [
                    {
                        "name": "invitation",
                        "in": "path",
                        "required": true,
                        "description": "The invitation ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`InvitationResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/InvitationResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
        "/service-tasks/{serviceTask}": {
            "get": {
                "operationId": "service-tasks.show",
                "summary": "Show a service task",
                "tags": [
                    "ServiceTask"
                ],
                "parameters": [
                    {
                        "name": "serviceTask",
                        "in": "path",
                        "required": true,
                        "description": "The service task ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ServiceTaskResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ServiceTaskResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            },
            "put": {
                "operationId": "service-tasks.update",
                "description": "Editing an interval never changes any vehicle's recorded service\nhistory; due dates move because they are derived.",
                "summary": "Update a service task",
                "tags": [
                    "ServiceTask"
                ],
                "parameters": [
                    {
                        "name": "serviceTask",
                        "in": "path",
                        "required": true,
                        "description": "The service task ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveServiceTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ServiceTaskResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ServiceTaskResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error.validation"
                    }
                }
            },
            "delete": {
                "operationId": "service-tasks.destroy",
                "description": "Only one no vehicle has history for (409 otherwise; deactivate it).",
                "summary": "Delete a service task",
                "tags": [
                    "ServiceTask"
                ],
                "parameters": [
                    {
                        "name": "serviceTask",
                        "in": "path",
                        "required": true,
                        "description": "The service task ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BayResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "branch_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "focus": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "capacity_hours_per_day": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "branch_id",
                    "name",
                    "focus",
                    "capacity_hours_per_day",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "title": "BayResource"
            },
...
        },
        "responses": {
            "Error.forbidden": {
                "description": "This action is unauthorized.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "forbidden"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable; never branch on it."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional machine-readable context."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
...
            "Error.not_found": {
                "description": "Not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "not_found"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable; never branch on it."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional machine-readable context."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}